+++ src/PPDS.Tools/Private/DataverseQueries.ps1
@@ -57,10 +57,8 @@ $script:DataverseImageTypeValues = @{
 }
 
 $script:ComponentType = @{
-    PluginType = 90
     PluginAssembly = 91
     SdkMessageProcessingStep = 92
-    SdkMessageProcessingStepImage = 93
 }
 
 function Get-PluginAssembly {
@@ -325,7 +323,7 @@ function Add-SolutionComponent {
         [Parameter(Mandatory = $true)]
         [string]$ComponentId,
         [Parameter(Mandatory = $true)]
-        [ValidateSet(90, 91, 92, 93)]
+        [ValidateSet(91, 92)]
         [int]$ComponentType,
         [Parameter()]
         [switch]$WhatIf
@@ -333,9 +331,7 @@ function Add-SolutionComponent {
 
     $componentTypeName = switch ($ComponentType) {
         91 { "Plugin Assembly" }
-        90 { "Plugin Type" }
         92 { "SDK Message Processing Step" }
-        93 { "SDK Message Processing Step Image" }
     }
 
     if ($WhatIf) {
@@ -343,7 +339,7 @@ function Add-SolutionComponent {
         return $true
     }
 
-    $addRequired = ($ComponentType -eq 90)
+    $addRequired = $false
 
     $body = @{
         ComponentId = $ComponentId